test(pr-maintenance): isolate rollup boundary cases

Copilot-Session: 0006555b-2f63-47a8-b61c-e8ead9b7f7b4
